+++ 윰_이유민/6주차 미션/미션 1/src/components/navbar/Navbar.style.js
@@ -1,13 +1,14 @@
-import styled from "styled-components";
-import { Link } from "react-router-dom";
+import styled from 'styled-components';
+import { Link } from 'react-router-dom';
 
 const Nav = styled.nav`
   height: 50px;
   display: flex;
   flex-direction: row;
+  align-items: center;
   justify-content: space-between;
   padding: 16px 14px;
-`
+`;
 
 const Logo = styled(Link)`
   color: red;
@@ -19,39 +20,6 @@ const Logo = styled(Link)`
   &:visited {
     color: red;
   }
-`
+`;
 
-const BtnContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 10px;
-
-  a {
-    width: 80px;
-    height: 30px;
-    border: none;
-    border-radius: 5px;
-    text-decoration: none;
-    text-align: center;
-    padding-top: 8px;
-  }
-`
-
-const Login = styled(Link)`
-  background-color: black;
-
-  &:hover {
-    background-color: white;
-    color: black;
-  }
-`
-
-const Signup = styled(Link)`
-  background-color: black;
-
-  &:hover {
-    background-color: red;
-  }
-`
-
-export { Nav, Logo, BtnContainer, Login, Signup }
+export { Nav, Logo };
